backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AI/LLM Integration
anthropic==0.18.0
google-generativeai==0.3.2
openai==1.10.0

# Vector Database
qdrant-client==1.7.3

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Text Processing
tiktoken==0.5.2
